+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added support for parsing a time stamp from a date string
+- Added support for serializing a time stamp for a given time zone
 - Added support for rudimentary time calculations
 - Split crate into three modules: `math`, `parse`, and `serde`
   - Each has to be opted-in with the corresponding feature flag